+++ rotkehlchen/chain/ethereum/modules/base_bridge/decoder.py
@@ -1,7 +1,6 @@
 from typing import TYPE_CHECKING, Any
 
-from rotkehlchen.assets.asset import AssetWithSymbol, EvmToken
-from rotkehlchen.chain.ethereum.utils import asset_normalized_value
+from rotkehlchen.assets.asset import AssetWithSymbol
 from rotkehlchen.chain.evm.decoding.constants import BASE_CPT_DETAILS
 from rotkehlchen.chain.evm.decoding.interfaces import DecoderInterface
 from rotkehlchen.chain.evm.decoding.structures import (
@@ -13,7 +12,6 @@ from rotkehlchen.chain.evm.decoding.types import CounterpartyDetails
 from rotkehlchen.chain.evm.decoding.utils import bridge_match_transfer, bridge_prepare_data
 from rotkehlchen.chain.evm.types import string_to_evm_address
 from rotkehlchen.constants.assets import A_ETH
-from rotkehlchen.constants.resolver import strethaddress_to_identifier
 from rotkehlchen.fval import FVal
 from rotkehlchen.types import ChainID, ChecksumEvmAddress
 from rotkehlchen.utils.misc import bytes_to_address
@@ -24,7 +22,6 @@ if TYPE_CHECKING:
     from rotkehlchen.user_messages import MessagesAggregator
 
 BRIDGE_ADDRESS = string_to_evm_address('0x49048044D57e1C92A77f79988d21Fa8fAF74E97e')
-TOKEN_PORTAL = string_to_evm_address('0x3154Cf16ccdb4C6d922629664174b904d80F2C35')
 TRANSACTION_DEPOSITED = b'\xb3\x815h\xd9\x99\x1f\xc9Q\x96\x1f\xcbLxH\x93WB@\xa2\x89%`M\t\xfcW|U\xbb|2'  # noqa: E501
 WITHDRAWAL_FINALIZED = b"\xdb\\vR\x85z\xa1c\xda\xad\xd6p\xe1\x16b\x8f\xb4.\x86\x9d\x8a\xc4%\x1e\xf8\x97\x1d\x9eW'\xdf\x1b"  # noqa: E501
 BRIDGE_TOKEN = b'\x7f\xf1&\xdb\x80$BK\xbf\xd9\x82n\x8a\xb8/\xf5\x916(\x9e\xa4@\xb0K9\xa0\xdf\x1b\x03\xb9\xca\xbf'  # noqa: E501
@@ -112,42 +109,11 @@ class BaseBridgeDecoder(DecoderInterface):
         )
         return DEFAULT_DECODING_OUTPUT
 
-    def _decode_bridge_tokens(self, context: DecoderContext) -> DecodingOutput:
-        """Decodes a bridging event for tokens. Either a deposit or a withdrawal."""
-        if context.tx_log.topics[0] not in {BRIDGE_TOKEN, WITHDRAWAL_TOKEN}:
-            # Make sure that we are decoding a supported event.
-            return DEFAULT_DECODING_OUTPUT
-
-        # Read information from event's topics & data
-        if context.tx_log.topics[0] == BRIDGE_TOKEN:
-            from_address = bytes_to_address(context.tx_log.topics[3])
-            to_address = bytes_to_address(context.tx_log.data[0:32])
-        else:
-            from_address = bytes_to_address(context.tx_log.data[0:32])
-            to_address = bytes_to_address(context.tx_log.topics[3])
-
-        raw_amount = int.from_bytes(context.tx_log.data[32:64])
-        asset = EvmToken(strethaddress_to_identifier(bytes_to_address(context.tx_log.topics[1])))
-        amount = asset_normalized_value(amount=raw_amount, asset=asset)
-
-        self._find_event(
-            context=context,
-            contract_address=TOKEN_PORTAL,
-            deposit_event=BRIDGE_TOKEN,
-            asset=asset,
-            from_address=from_address,
-            to_address=to_address,
-            amount=amount,
-        )
-
-        return DEFAULT_DECODING_OUTPUT
-
     # -- DecoderInterface methods
 
     def addresses_to_decoders(self) -> dict[ChecksumEvmAddress, tuple[Any, ...]]:
         return {
             BRIDGE_ADDRESS: (self._decode_bridge_eth,),
-            TOKEN_PORTAL: (self._decode_bridge_tokens,),
         }
 
     @staticmethod