+++ rotkehlchen/tests/unit/decoders/test_superchain_bridge.py
@@ -2,15 +2,13 @@ import pytest
 
 from rotkehlchen.accounting.structures.balance import Balance
 from rotkehlchen.assets.asset import Asset
-from rotkehlchen.chain.ethereum.modules.optimism_bridge.decoder import (
-    OPTIMISM_L1_ESCROW,
-    OPTIMISM_PORTAL_ADDRESS,
-)
+from rotkehlchen.chain.base.constants import CPT_BASE
+from rotkehlchen.chain.ethereum.modules.superchain_bridge.op.decoder import OPTIMISM_PORTAL_ADDRESS
 from rotkehlchen.chain.evm.constants import ZERO_ADDRESS
 from rotkehlchen.chain.evm.decoding.constants import CPT_GAS
 from rotkehlchen.chain.evm.types import string_to_evm_address
 from rotkehlchen.chain.optimism.constants import CPT_OPTIMISM
-from rotkehlchen.constants.assets import A_DAI, A_ETH, A_OPTIMISM_ETH
+from rotkehlchen.constants.assets import A_ETH, A_OPTIMISM_ETH
 from rotkehlchen.fval import FVal
 from rotkehlchen.history.events.structures.evm_event import EvmEvent
 from rotkehlchen.history.events.structures.types import HistoryEventSubType, HistoryEventType
@@ -371,42 +369,6 @@ def test_prove_withdrawal(ethereum_inquirer, ethereum_accounts):
 
 
 @pytest.mark.vcr(filter_query_parameters=['apikey'])
-@pytest.mark.parametrize('ethereum_accounts', [['0x74eEFF1A4D24C9Cd48FF9D8Dabd9Db0120b9Caf6']])
-def test_deposit_dai_on_ethereum(ethereum_inquirer, ethereum_accounts):
-    evmhash = deserialize_evm_tx_hash('0xad659c207e54b3fdb2124e38b8cfa673851c49feb97115f45ba396f80a4fd35b')  # noqa: E501
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=evmhash)
-    user_address, timestamp, gas_amount, deposit_amount = ethereum_accounts[0], TimestampMS(1723807967000), '0.0006873796276356', '97.27'  # noqa: E501
-    assert events == [
-        EvmEvent(
-            tx_hash=evmhash,
-            sequence_index=0,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.SPEND,
-            event_subtype=HistoryEventSubType.FEE,
-            asset=A_ETH,
-            balance=Balance(amount=FVal(gas_amount)),
-            location_label=user_address,
-            notes=f'Burn {gas_amount} ETH for gas',
-            counterparty=CPT_GAS,
-        ), EvmEvent(
-            tx_hash=evmhash,
-            sequence_index=227,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.DEPOSIT,
-            event_subtype=HistoryEventSubType.BRIDGE,
-            asset=A_DAI,
-            balance=Balance(amount=FVal(deposit_amount)),
-            location_label=user_address,
-            notes=f'Bridge {deposit_amount} DAI from Ethereum to Optimism via Optimism bridge',
-            counterparty=CPT_OPTIMISM,
-            address=OPTIMISM_L1_ESCROW,
-        ),
-    ]
-
-
-@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('optimism_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
 def test_deposit_dai_on_optimism(optimism_inquirer, optimism_accounts):
     evmhash = deserialize_evm_tx_hash('0xd566bc86bdc88cf811fee5bbf5f233cb8add71642fdf547866f4054c4362a922')  # noqa: E501
@@ -443,42 +405,6 @@ def test_deposit_dai_on_optimism(optimism_inquirer, optimism_accounts):
 
 
 @pytest.mark.vcr(filter_query_parameters=['apikey'])
-@pytest.mark.parametrize('ethereum_accounts', [['0x9531C059098e3d194fF87FebB587aB07B30B1306']])
-def test_withdraw_dai_on_ethereum(ethereum_inquirer, ethereum_accounts):
-    evmhash = deserialize_evm_tx_hash('0x3e958996b7c4b2466ca9521d4c688ff4817c28d66bb3f52ec1174c6d0f1c44db')  # noqa: E501
-    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=evmhash)
-    user_address, timestamp, gas_amount, receive_amount = ethereum_accounts[0], TimestampMS(1671053711000), '0.0097257501', '106317.949168317256453804'  # noqa: E501
-    assert events == [
-        EvmEvent(
-            tx_hash=evmhash,
-            sequence_index=0,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.SPEND,
-            event_subtype=HistoryEventSubType.FEE,
-            asset=A_ETH,
-            balance=Balance(amount=FVal(gas_amount)),
-            location_label=user_address,
-            notes=f'Burn {gas_amount} ETH for gas',
-            counterparty=CPT_GAS,
-        ), EvmEvent(
-            tx_hash=evmhash,
-            sequence_index=876,
-            timestamp=timestamp,
-            location=Location.ETHEREUM,
-            event_type=HistoryEventType.WITHDRAWAL,
-            event_subtype=HistoryEventSubType.BRIDGE,
-            asset=A_DAI,
-            balance=Balance(amount=FVal(receive_amount)),
-            location_label=user_address,
-            notes=f'Bridge {receive_amount} DAI from Optimism to Ethereum via Optimism bridge',
-            counterparty=CPT_OPTIMISM,
-            address=OPTIMISM_L1_ESCROW,
-        ),
-    ]
-
-
-@pytest.mark.vcr(filter_query_parameters=['apikey'])
 @pytest.mark.parametrize('optimism_accounts', [['0x301605C95acbED7A1fD9C2c0DeEe964e2AFBd0C3']])
 def test_withdraw_dai_on_optimism(optimism_inquirer, optimism_accounts):
     evmhash = deserialize_evm_tx_hash('0x785d75b9a5c93b2ad47b662f8f98f9c63d31cd629497bcbd846c57a70f76366e')  # noqa: E501
@@ -500,3 +426,200 @@ def test_withdraw_dai_on_optimism(optimism_inquirer, optimism_accounts):
             address=ZERO_ADDRESS,
         ),
     ]
+
+
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
+@pytest.mark.parametrize('ethereum_accounts', [['0x6D376A5863324489BD65AB2cc8E24d6cE3775848']])
+def test_deposit_eth_ethereum_to_base_bridge(ethereum_inquirer, ethereum_accounts):
+    evmhash = deserialize_evm_tx_hash('0xe5e26e70a257ca733042d8cbbc4764737d115016892e440ad6cea43933a655d7')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=evmhash)
+    user_address, timestamp, deposit_amount = ethereum_accounts[0], TimestampMS(1730128127000), '2'
+    expected_events = [EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=0,
+            timestamp=timestamp,
+            location=Location.ETHEREUM,
+            event_type=HistoryEventType.SPEND,
+            event_subtype=HistoryEventSubType.FEE,
+            asset=A_ETH,
+            balance=Balance(amount=FVal('0.001994757244065416')),
+            location_label=user_address,
+            notes='Burn 0.001994757244065416 ETH for gas',
+            counterparty=CPT_GAS,
+        ), EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=1,
+            timestamp=timestamp,
+            location=Location.ETHEREUM,
+            event_type=HistoryEventType.DEPOSIT,
+            event_subtype=HistoryEventSubType.BRIDGE,
+            asset=A_ETH,
+            balance=Balance(amount=FVal(deposit_amount)),
+            location_label=user_address,
+            notes=f'Bridge {deposit_amount} ETH from Ethereum to Base via Base bridge',
+            counterparty=CPT_BASE,
+            address=string_to_evm_address('0x3154Cf16ccdb4C6d922629664174b904d80F2C35'),
+        ),
+    ]
+    assert events == expected_events
+
+
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
+@pytest.mark.parametrize('ethereum_accounts', [['0x150f4a772a640BeD3a33E5919D1E3f4fc8dE2cd0']])
+def test_deposit_erc20_ethereum_to_base_bridge(ethereum_inquirer, ethereum_accounts):
+    evmhash = deserialize_evm_tx_hash('0x58160ce93b20fa0ec513dd956325ef8e11f9478af21573bb3986b3fb167f475e')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=ethereum_inquirer, tx_hash=evmhash)
+    gas_amount, deposit_amount = '0.002095833929196665', '15828.918676384591706627'
+    user_address, timestamp = ethereum_accounts[0], TimestampMS(1730111759000)
+    expected_events = [EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=0,
+            timestamp=timestamp,
+            location=Location.ETHEREUM,
+            event_type=HistoryEventType.SPEND,
+            event_subtype=HistoryEventSubType.FEE,
+            asset=A_ETH,
+            balance=Balance(amount=FVal(gas_amount)),
+            location_label=user_address,
+            notes=f'Burn {gas_amount} ETH for gas',
+            counterparty=CPT_GAS,
+        ), EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=148,
+            timestamp=timestamp,
+            location=Location.ETHEREUM,
+            event_type=HistoryEventType.DEPOSIT,
+            event_subtype=HistoryEventSubType.BRIDGE,
+            asset=Asset('eip155:1/erc20:0x44ff8620b8cA30902395A7bD3F2407e1A091BF73'),
+            balance=Balance(amount=FVal(deposit_amount)),
+            location_label=user_address,
+            notes=f'Bridge {deposit_amount} VIRTUAL from Ethereum to Base via Base bridge',
+            counterparty=CPT_BASE,
+            address=string_to_evm_address('0x3154Cf16ccdb4C6d922629664174b904d80F2C35'),
+        ),
+    ]
+    assert events == expected_events
+
+
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
+@pytest.mark.parametrize('base_accounts', [['0xd34Ec2202b56261b6d7586a752E37a36818c0538']])
+def test_receive_eth_ethereum_to_base_bridge(base_inquirer, base_accounts):
+    evmhash = deserialize_evm_tx_hash('0x924167064e29ecb68ea9586811b63799b6154158a6607ac9fcf9530667446a5c')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=base_inquirer, tx_hash=evmhash)
+    amount = '0.83'
+    user_address, timestamp = base_accounts[0], TimestampMS(1730133657000)
+    expected_events = [EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=0,
+            timestamp=timestamp,
+            location=Location.BASE,
+            event_type=HistoryEventType.WITHDRAWAL,
+            event_subtype=HistoryEventSubType.BRIDGE,
+            asset=A_ETH,
+            balance=Balance(amount=FVal(amount)),
+            location_label=user_address,
+            notes=f'Bridge {amount} ETH from Ethereum to Base via Base bridge',
+            counterparty=CPT_BASE,
+            address=string_to_evm_address('0x4200000000000000000000000000000000000010'),
+        ),
+    ]
+    assert events == expected_events
+
+
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
+@pytest.mark.parametrize('base_accounts', [['0x1218d6396dC67eC0FFBEBDF049C865B83636EddA']])
+def test_receive_erc20_ethereum_to_base_bridge(base_inquirer, base_accounts):
+    evmhash = deserialize_evm_tx_hash('0xf731b5c5dc53117413b880d0c65501cebe50c78ba3621eca9344747c95b83357')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=base_inquirer, tx_hash=evmhash)
+    amount = '4900031.556715611515583027'
+    user_address, timestamp = base_accounts[0], TimestampMS(1730133393000)
+    expected_events = [EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=0,
+            timestamp=timestamp,
+            location=Location.BASE,
+            event_type=HistoryEventType.WITHDRAWAL,
+            event_subtype=HistoryEventSubType.BRIDGE,
+            asset=Asset('eip155:8453/erc20:0x7Cea109FC3516eD1248ae9AA67B5a352cF74075e'),
+            balance=Balance(amount=FVal(amount)),
+            location_label=user_address,
+            notes=f'Bridge {amount} NOVA from Ethereum to Base via Base bridge',
+            counterparty=CPT_BASE,
+            address=ZERO_ADDRESS,
+        ),
+    ]
+    assert events == expected_events
+
+
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
+@pytest.mark.parametrize('base_accounts', [['0x6730b1Df17E50217777EeE475E34815964e3BFb2']])
+def test_withdraw_eth_base_to_ethereum_bridge(base_inquirer, base_accounts):
+    evmhash = deserialize_evm_tx_hash('0xe451ca095dd9d48f6558a226fc6cc9b28d19f39080545db63b8ba9410fe3df3e')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=base_inquirer, tx_hash=evmhash)
+    gas_amount, deposit_amount = '0.000000442236014244', '133.00839331231264871'
+    user_address = base_accounts[0]
+    expected_events = [EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=0,
+            timestamp=TimestampMS(1729279481000),
+            location=Location.BASE,
+            event_type=HistoryEventType.SPEND,
+            event_subtype=HistoryEventSubType.FEE,
+            asset=A_ETH,
+            balance=Balance(amount=FVal(gas_amount)),
+            location_label=user_address,
+            notes=f'Burn {gas_amount} ETH for gas',
+            counterparty=CPT_GAS,
+        ), EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=1,
+            timestamp=TimestampMS(1729279481000),
+            location=Location.BASE,
+            event_type=HistoryEventType.DEPOSIT,
+            event_subtype=HistoryEventSubType.BRIDGE,
+            asset=A_ETH,
+            balance=Balance(amount=FVal(deposit_amount)),
+            location_label=user_address,
+            notes=f'Bridge {deposit_amount} ETH from Base to Ethereum via Base bridge',
+            counterparty=CPT_BASE,
+            address=string_to_evm_address('0x4200000000000000000000000000000000000010'),
+        ),
+    ]
+    assert events == expected_events
+
+
+@pytest.mark.vcr(filter_query_parameters=['apikey'])
+@pytest.mark.parametrize('base_accounts', [['0x779e4b47c3Dea5689233821dFcf429E0485eF116']])
+def test_withdraw_erc20_base_to_ethereum_bridge(base_inquirer, base_accounts):
+    evmhash = deserialize_evm_tx_hash('0xbe4e54e77cb700f2755b236d7823295129c4d5e22fdc87df8058274bc0fefab1')  # noqa: E501
+    events, _ = get_decoded_events_of_transaction(evm_inquirer=base_inquirer, tx_hash=evmhash)
+    gas_amount, deposit_amount = '0.000029210771559657', '49352.072702'
+    user_address = base_accounts[0]
+    expected_events = [EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=0,
+            timestamp=TimestampMS(1729293119000),
+            location=Location.BASE,
+            event_type=HistoryEventType.SPEND,
+            event_subtype=HistoryEventSubType.FEE,
+            asset=A_ETH,
+            balance=Balance(amount=FVal(gas_amount)),
+            location_label=user_address,
+            notes=f'Burn {gas_amount} ETH for gas',
+            counterparty=CPT_GAS,
+        ), EvmEvent(
+            tx_hash=evmhash,
+            sequence_index=79,
+            timestamp=TimestampMS(1729293119000),
+            location=Location.BASE,
+            event_type=HistoryEventType.DEPOSIT,
+            event_subtype=HistoryEventSubType.BRIDGE,
+            asset=Asset('eip155:8453/erc20:0xd9aAEc86B65D86f6A7B5B1b0c42FFA531710b6CA'),
+            balance=Balance(amount=FVal(deposit_amount)),
+            location_label=user_address,
+            notes=f'Bridge {deposit_amount} USDbC from Base to Ethereum via Base bridge',
+            counterparty=CPT_BASE,
+            address=ZERO_ADDRESS,
+        ),
+    ]
+    assert events == expected_events